changed the way action is cast to int in model.py

+++ dqn_agent.py
@@ -76,9 +76,9 @@ class Agent():
 
         # Epsilon-greedy action selection
         if random.random() > eps:
-            return np.int32(np.argmax(action_values.cpu().data.numpy())) #int32 here makes unity env happy
+            return np.argmax(action_values.cpu().data.numpy()).astype('int') #int32 here makes unity env happy
         else:
-            return random.choice(np.arange(self.action_size))
+            return random.choice(np.arange(self.action_size)).astype('int')
 
     def learn(self, experiences, gamma):
         """Update value parameters using given batch of experience tuples.
